recommendation-service/requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3
pandas==2.0.0
scikit-learn==1.2.2
numpy==1.24.3
py-eureka-client==0.11.8
python-dotenv==1.0.0 